refactor: css style files organisation
