+++ src/web/src/app.tsx
@@ -2,10 +2,11 @@
 
 import { Toaster } from "react-hot-toast";
 import { ReactQuery } from "./lib/react-query.tsx";
-import { IndexView } from "./view/index-view.tsx";
-import { Footer } from "./ui/footer.tsx";
+import { IndexView } from "./ui/view/index/index-view.tsx";
+import { Footer } from "./ui/layout/footer/footer.tsx";
 import { Container } from "./ui/container.tsx";
 import { LinguiProvider } from "./i18n/lingui-provider.tsx";
+import { appConfig } from "./config/app-config.ts";
 import "./ui/app.css";
 
 function App() {
@@ -14,7 +15,7 @@ function App() {
       <ReactQuery>
         <Container>
           <IndexView />
-          <Footer />
+          <Footer gitHubUrl={appConfig.githubUrl} />
           <Toaster />
         </Container>
       </ReactQuery>
