+++ src/web/src/config/app-config.ts
@@ -1,0 +1,3 @@
+export const appConfig = {
+  githubUrl: "https://github.com/landsman/music",
+};
