+++ src/web/src/ui/activity/_activity.css
@@ -1,0 +1,2 @@
+@import "loader/loader.css";
+@import "pull-to-refresh/pull-to-refresh.css";
