+++ src/web/src/ui/activity/loader/loader.css
