+++ src/web/src/ui/activity/pull-to-refresh/pull-to-refresh.css
