+++ src/web/src/ui/app.css
@@ -1,9 +1,12 @@
-@import "./styles/reset.css";
-@import "./styles/layout.css";
+/* layout */
+@import "layout/_layout.css";
 
-/* components */
-@import "./activity/loader.module.css";
-@import "feed/list/feed.module.css";
+/* general components */
+@import "activity/_activity.css";
+@import "form/_form.css";
 
 /* views */
-@import "../view/index-view.module.css";
+@import "view/_view.css";
+
+/* features */
+@import "feature/_feature.css";
