+++ src/web/src/ui/feature/_feature.css
@@ -1,0 +1,1 @@
+@import "feed/feed.css";
