+++ src/web/src/ui/feature/feed/item/feed-item.tsx
@@ -3,13 +3,13 @@ import { useLingui } from "@lingui/react/macro";
 import {
   localizeDateTimeBrowser,
   localizeRelativeTimeBrowser,
-} from "../../../lib/localize-date.ts";
+} from "../../../../lib/localize-date.ts";
 
 import { Marquee } from "../track/track.tsx";
-import { User } from "../../user/user.tsx";
+import { User } from "../../../user/user.tsx";
 import { Album } from "../album.tsx";
 import { Artist } from "../artist.tsx";
-import { usePlaySound } from "../../../lib/use-play-sound.ts";
+import { usePlaySound } from "../../../../lib/use-play-sound.ts";
 
 interface Props {
   index: number;
