+++ src/web/src/ui/form/_form.css
@@ -1,0 +1,1 @@
+@import "select.css";
