+++ src/web/src/ui/form/select.css
@@ -1,0 +1,21 @@
+select {
+  padding: 0.4rem 0.6rem;
+  border-radius: 4px;
+  border: 1px solid
+    color-mix(in oklab, var(--color-neutral-700) 75%, transparent);
+  background-color: transparent;
+  color: inherit;
+  font-size: 14px;
+  cursor: pointer;
+
+  &:focus {
+    outline: none;
+    border-color: #007bff;
+    box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
+  }
+
+  option {
+    background-color: #242424;
+    color: rgba(255, 255, 255, 0.87);
+  }
+}
