--- src/web/src/ui/header/header.css
@@ -1,33 +1,0 @@
-.header-top {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.header-right {
-  display: flex;
-  align-items: center;
-}
-
-.user-filter select {
-  padding: 0.4rem 0.6rem;
-  border-radius: 4px;
-  border: 1px solid
-    color-mix(in oklab, var(--color-neutral-700) 75%, transparent);
-  background-color: transparent;
-  color: inherit;
-  font-size: 14px;
-  cursor: pointer;
-}
-
-.user-filter select:focus {
-  outline: none;
-  border-color: #007bff;
-  box-shadow: 0 0 0 2px rgba(0, 123, 255, 0.25);
-}
-
-.user-filter select option {
-  background-color: #242424;
-  color: rgba(255, 255, 255, 0.87);
-}
