+++ src/web/src/ui/layout/footer/footer.css
@@ -1,16 +1,3 @@
-#root {
-  width: 1280px;
-  max-width: 100%;
-  overflow: hidden;
-  margin: 0 auto;
-  padding: 1rem;
-}
-
-.container {
-  max-width: 600px;
-  margin: 0 auto;
-}
-
 .footer {
   display: flex;
   flex-direction: row;
