+++ src/web/src/ui/layout/global/layout.css
@@ -1,0 +1,12 @@
+#root {
+  width: 1280px;
+  max-width: 100%;
+  overflow: hidden;
+  margin: 0 auto;
+  padding: 1rem;
+}
+
+.container {
+  max-width: 600px;
+  margin: 0 auto;
+}
