+++ src/web/src/ui/layout/header/header.css
@@ -1,0 +1,11 @@
+.header-top {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+}
+
+.header-right {
+  display: flex;
+  align-items: center;
+}
