+++ src/web/src/ui/view/_view.css
@@ -1,0 +1,1 @@
+@import "index/index-view.css";
