+++ src/web/src/ui/view/index/index-view.css
